move test for ingredient count options
